requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

typing-extensions==3.7.4.3
Mako==1.1.4
redis==3.2.0
django==2.2.6
# blueapps==3.3.3
gunicorn==19.6.0
django-filter==2.4.0
djangorestframework==3.11.0
django-cors-headers==3.5.0
elasticsearch==7.10.0
bamboo-pipeline==3.1.6
bamboo_engine==1.0.8
# django-celery==3.3.0
anyjson==0.3.3
django-apscheduler==0.5.2
urllib3==1.25.11
mysqlclient==2.0.3
pytz==2019.3
django-proxy==1.2.1
